backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
djangorestframework-simplejwt==5.3.0
stripe==7.8.0
requests==2.31.0
python-telegram-bot==20.7
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
